// pos/src/app/models/m-user-details.ts
export class MUserDetails {
    constructor(){}

    id?: number;
    userCode?: string;
    userName?: string;
    posCode?: string;
    storeCode?: string;
    password?: string;
    employeeID?: number;
    roleID?: number;
    priceListID?: number;
    priceListCode?: string;
    countryID?: number;
    storeID?: number;
    posid?: number;
    currencyID?: number;
    storeName?: string;
    countryName?: string;
    roleName?: string;
    employeeName?: string;
    storeGroupName?: string;
    brandID?: string;
    storeGroupID?: number;
    allowNewRow?: boolean;
    employeeImage?: string;
    printerDeviceName?: string;
    posName?: string;
    currencySymbol?: string;
    baseCurrency?: string;
    managerOverrideID?: number;
    retailID?: number;
    managerOverrideName?: string;
    retailName?: string;
    decimalPlaces?: number;
    nearByRoundOff?: number;
    passwordReset?: boolean;
    currentPassword?: string;
    confirmPassword?: string;
    isLoggedIn?: boolean;
    poleDisplayPort?: string;
    displayLineMsgOne?: string;
    displayLineMsgTwo?: string;
    printCount?: number;
    exchangePrintCount?: number;
    saleReturnPrintCount?: number;
    countryCode?: string;
    employeeCode?: string;
    toMailID?: string;
    ccMailID?: string;
    managerOverrideCode?: string;
    retailSettingCode?: string;
    roleCode?: string;
    currencyCode?: string;
    posTitle?:string;
    allowStockEdit?: boolean;
    mobileUser?: boolean;
    promotionRoundOff?: number;
    taxID?: number;
    taxCode?: string;
    taxPercentage?: number;
    storeFooter?:string;
    storeImage?:string;

    // active?: boolean;
    // scn?: number;
    // createBy?: number;
    // createOn?: Date;
    // updateBy?: number;
    // updateOn?: Date;
    // createdByUserName?: string;
    // updatedByUserName?: string;
    // isStoreSync?: boolean;
    // isCountrySync?: boolean;
    // appVersion?: string;
    // isServerSync?: boolean;
}